requirements.txt
streamlit==1.41.1
langchain-groq==0.2.3
langchain-community==0.3.15
langchain==0.3.15
langchain-core==0.3.31
langchain-huggingface==0.1.2
chromadb==0.6.3
python-dotenv==1.0.1
huggingface-hub==0.27.1
tiktoken==0.8.0
beautifulsoup4==4.12.3
faiss-cpu==1.9.0.post1


